sisTOPOGRAFIA/py_engine/requirements-lock.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
astroid==4.0.4
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.4
crcmod==1.7
cryptography==46.0.5
curl_cffi==0.13.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
ecdsa==0.19.1
et_xmlfile==2.0.0
ezdxf==1.4.3
fastkml==1.4.0
flake8==7.3.0
fonttools==4.61.1
fpdf2==2.8.5
frozendict==2.4.7
future==1.0.0
geocoder==1.38.1
geopandas==1.1.2
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
joblib==1.5.3
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.8
mccabe==0.7.0
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.2
openpyxl==3.1.5
osmnx==2.0.7
packaging==26.0
pandas==3.0.0
passlib==1.7.4
peewee==3.19.0
pefile==2024.8.26
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
protobuf==6.33.5
pyasn1==0.6.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
pygeoif==1.6.0
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.0
pylint==4.0.4
pyogrio==0.12.1
pyparsing==3.3.2
pyperclip==1.11.0
pyproj==3.7.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
ratelim==0.1.6
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==82.0.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
threadpoolctl==3.6.0
tkintermapview==1.29
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.1.0
